# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WGPanel API
  description: |
    WGPanel manages WireGuard nodes and peer ("account") provisioning behind a
    separate Telegram sales bot. This panel deliberately has **no concept of
    payments, pricing, or plans** — the bot owns all of that and talks to this
    API using a scoped, HMAC-signed API key. Admin/operator users manage the
    same resources through the same endpoints using a JWT session instead.

    ## Authentication

    Every endpoint under `/api/v1/*` (except `POST /api/v1/auth/login`,
    `POST /api/v1/auth/refresh`, and `POST /api/v1/nodes/join`) requires
    **one** of:

    - **Admin session** — `Authorization: Bearer <access_token>` from
      `POST /api/v1/auth/login`. Access tokens are short-lived (15 minutes);
      exchange the refresh token via `POST /api/v1/auth/refresh` before it
      expires. Admin sessions are additionally subject to role checks
      (`support` < `operator` < `super_admin`) documented per-operation below.
    - **Bot/reseller API key** — three headers together:
      `X-API-Key` (the key ID), `X-Timestamp` (unix seconds), and
      `X-Signature` (hex HMAC-SHA256). The signed payload is exactly:

      ```
      {METHOD}\n{URL_PATH}\n{TIMESTAMP}\n{RAW_BODY}
      ```

      `METHOD` is the uppercase HTTP verb, `URL_PATH` is the request path
      with no query string, `TIMESTAMP` is the literal `X-Timestamp` header
      value (not re-serialized), and `RAW_BODY` is the exact request body
...

...

...

...

...

...

...
servers:
...
paths:
...
  /agent/register:
    servers:
      - url: https://node-agent.internal:9090
    post:
      tags: [Node Agent]
      summary: Redeem a join token for an mTLS client certificate
      description: >-
        Called by `cmd/agent` on first startup. No client certificate is
        required for this call (it's how the node gets one) - the `join_token`
        field is the sole credential, single-use, redeemed atomically. The
        returned certificate's Common Name is always the server-assigned node
        ID; the CSR's own requested Subject/CN is ignored, so a compromised or
        buggy agent can never claim a different node's identity.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [join_token, csr_pem]
              properties:
                join_token: { type: string }
                csr_pem:
                  type: string
                  description: PEM-encoded PKCS#10 certificate signing request, generated locally by the agent (its private key never leaves the node).
                wg_public_key:
                  type: string
                  description: This node's own WireGuard server public key, if install-node.sh already created a real wg0 interface. Optional.
      responses:
        "201":
          description: Registered. Store the certificate and CA alongside the agent's private key for all future mTLS calls.
          content:
            application/json:
              schema:
                type: object
                properties:
                  node_id: { type: string, format: uuid }
                  certificate_pem: { type: string }
                  ca_certificate_pem: { type: string }
        "400":
          description: "`invalid_request` (missing fields) or `invalid_csr` (CSR could not be parsed/signed)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          description: "`invalid_token` - join token invalid, expired, or already used."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "500": { $ref: "#/components/responses/InternalError" }
  /agent/heartbeat:
    servers:
      - url: https://node-agent.internal:9090
    post:
      tags: [Node Agent]
      summary: Report liveness and peer count
      description: >-
        Called every 10 seconds by `cmd/agent`. **Requires a client certificate**
        issued by a prior `/agent/register` call - enforced at the TLS layer
        (connection-level) and again in-handler (the cert's Subject CN is
        looked up as the node ID, and its fingerprint must match what was
        pinned at registration, so a stale-but-still-CA-signed cert from a
        node that re-registered is correctly rejected). This requirement
        cannot be expressed as an OpenAPI security scheme in 3.0 - treat it as
        a hard transport-level precondition, not optional.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                peer_count: { type: integer }
                load_avg: { type: number }
              description: Informational only - malformed/absent bodies are tolerated, not rejected.
      responses:
        "200":
          description: Heartbeat recorded; node status set to `online`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        "401":
          description: "`client_certificate_required` (no client cert presented) or `certificate_superseded` (cert's fingerprint no longer matches this node's pinned identity)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "404":
          description: "`node_not_found` - certificate's Common Name doesn't match any known node."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin session access token from POST /api/v1/auth/login (15 minute TTL).
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Bot/reseller auth. Must be combined with X-Timestamp and X-Signature
        headers (not separately modelable as OpenAPI security schemes since
        OpenAPI has no concept of "N headers form one signed credential") -
        see the top-level description for the exact signing scheme.
    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: Must equal the INTERNAL_API_TOKEN env var. Gates every /internal/* route.
  responses:
...
    InternalError:
      description: "`internal_error` - unexpected server-side failure."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: >-
                Stable machine-readable identifier. Known values include:
                invalid_request, invalid_csr, invalid_credentials,
                invalid_refresh_token, invalid_token, client_certificate_required,
                certificate_superseded, unauthorized, invalid_timestamp,
                stale_timestamp, forbidden, node_group_not_allowed, node_not_found,
                account_not_found, api_key_not_found, node_not_registered,
                node_capacity_exceeded, no_available_node, external_ref_taken,
                node_missing_public_key, username_taken, key_id_collision,
                node_name_taken, not_ready, internal_error.
            message:
              type: string
              description: Human-readable; may change wording - do not match against this.
      required: [error]
...